fix(cli): move ping unit test to an external module file
The perfectionist::unit_test_file_layout dylint lint rejects inline
`#[cfg(test)] mod tests { ... }` blocks. Move the redact_and_sanitize
test into src/cli_args/ping/tests.rs and declare it with an external
`mod tests;`, matching the convention used by the other cli_args modules.
